+++ auto_qc/util/functional.py
@@ -1,22 +1,6 @@
 import typing
 from functools import reduce
 
-T = typing.TypeVar("T")
-
-
-def identity(x: T) -> T:
-    """
-    The identity function.
-    """
-    return x
-
-
-def empty_list(*args: typing.Any) -> list[typing.Any]:
-    """
-    Returns an empty list, whatever the arguments.
-    """
-    return []
-
 
 def flatten(n: list[typing.Any]) -> list[typing.Any]:
     def _f(acc: list[typing.Any], x: typing.Any) -> list[typing.Any]:
@@ -37,22 +21,3 @@ def get_in(data: typing.Any, keys: typing.Iterable[str]) -> typing.Any:
             return None
         data = data[key]
     return data
-
-
-def recursive_apply(
-    list_func: typing.Callable[[list[typing.Any]], typing.Any],
-    atom_func: typing.Callable[[typing.Any], typing.Any] = identity,
-) -> typing.Callable[[typing.Any], typing.Any]:
-    """
-    Creates a function which applies either of the two given functions: the first
-    to a list, and the second to atoms within a list. Used to walk over deeply
-    nested s-expressions.
-    """
-
-    def _f(x: typing.Any) -> typing.Any:
-        if isinstance(x, list):
-            return list_func(x)
-        else:
-            return atom_func(x)
-
-    return _f